Merge from Moya

+++ Rakefile
@@ -14,8 +14,8 @@ def safe_sh(command)
   end
 end
 
-def workspace
-  return 'Demo.xcworkspace'
+def moya_project
+  return 'Moya.xcodeproj'
 end
 
 def configuration
@@ -24,7 +24,7 @@ end
 
 def targets
   return [
-    # :macos, # Note: we're experiencing macOS build problems on circle, commenting out.
+    :macos,
     :tvos,
     :ios
   ]
@@ -32,9 +32,9 @@ end
 
 def schemes
   return {
-    ios: 'Demo',
-    macos: 'MoyaTests-Mac',
-    tvos: 'MoyaTests-tvOS'
+    ios: 'MoyaTests',
+    macos: 'MoyaTests',
+    tvos: 'MoyaTests'
   }
 end
 
@@ -46,54 +46,74 @@ def sdks
   }
 end
 
-# See: https://discuss.circleci.com/t/multiple-ios-simulators-available-for-single-os-device-configurations/7854
 def devices
   return {
-    ios: ENV['CIRCLECI'] ? "id='F08BA729-6AD2-42DF-A210-34DC8D990011'" : "name='iPhone 6s'",
-    macos: "arch='x86_64'",
-    tvos: ENV['CIRCLECI'] ? "id='48B0E1AB-F5EB-40FB-9372-A16B93349B12'" : "name='Apple TV 1080p'"
+    ios: "OS=#{device_os[:ios]},name=#{device_names[:ios]}",
+    macos: "arch=x86_64",
+    tvos: "OS=#{device_os[:tvos]},name=#{device_names[:tvos]}"
   }
 end
 
-# See: https://discuss.circleci.com/t/xcode-exit-code-65/4284/13
-def uuids
+def device_names
   return {
-    ios: 'F08BA729-6AD2-42DF-A210-34DC8D990011', # iPhone 6s
-    tvos: '48B0E1AB-F5EB-40FB-9372-A16B93349B12' # Apple TV 1080p
+    ios: "iPhone 6s",
+    tvos: "Apple TV 4K (at 1080p)"
   }
 end
 
-def open_simulator_and_sleep(uuid)
-  return if uuid.nil? # Don't need a sleep on macOS because it runs first.
-  sh "xcrun instruments -w '#{uuid}' || sleep 15"
+def device_os
+  return {
+    ios: "11.2",
+    tvos: "11.2"
+  }
 end
 
-def xcodebuild_in_demo_dir(tasks, platform, xcprety_args: '')
+def open_simulator_and_sleep(platform)
+  return if platform == :macos # Don't need a sleep on macOS because it runs first.
+  sh "xcrun instruments -w '#{device_names[platform]} (#{device_os[platform]})' || sleep 15"
+end
+
+def xcodebuild(tasks, platform, xcprety_args: '')
   sdk = sdks[platform]
   scheme = schemes[platform]
   destination = devices[platform]
 
-  Dir.chdir('Demo') do
-    open_simulator_and_sleep(uuids[platform])
-    safe_sh "set -o pipefail && xcodebuild -workspace '#{workspace}' -scheme '#{scheme}' -configuration '#{configuration}' -sdk #{sdk} -destination #{destination} #{tasks} | bundle exec xcpretty -c #{xcprety_args}"
-  end
+  open_simulator_and_sleep(platform)
+  safe_sh "set -o pipefail && xcodebuild -project '#{moya_project}' -scheme '#{scheme}' -configuration '#{configuration}' -sdk #{sdk} -destination '#{destination}' #{tasks} | bundle exec xcpretty -c #{xcprety_args}"
+end
+
+def xcodebuild_example(tasks, xcprety_args: '')
+  platform = :ios
+  sdk = sdks[platform]
+  destination = devices[platform]
+  demo_project = 'Moya.xcodeproj'
+  demo_scheme = 'Basic'
+
+  open_simulator_and_sleep(platform)
+  safe_sh "set -o pipefail && xcodebuild -project '#{demo_project}' -scheme '#{demo_scheme}' -configuration '#{configuration}' -sdk #{sdk} -destination '#{destination}' #{tasks} | bundle exec xcpretty -c #{xcprety_args}"
+end
+
+desc 'Build Moya.'
+task :build do
+  xcodebuild 'build', :ios
 end
 
 desc 'Build the Demo app.'
-task :build do
-  xcodebuild_in_demo_dir 'build', :ios
+task :build_example do
+  xcodebuild_example 'build'
 end
 
 desc 'Clean build directory.'
 task :clean do
-  xcodebuild_in_demo_dir 'clean', :ios
+  xcodebuild 'clean', :ios
 end
 
 desc 'Build, then run all tests.'
 task :test do
   targets.map do |platform|
     puts "Testing on #{platform}."
-    xcodebuild_in_demo_dir 'build test', platform, xcprety_args: '--test'
+    xcodebuild 'build test', platform, xcprety_args: '--test'
+    next unless platform == :mac
     sh "killall Simulator"
   end
 end
@@ -102,18 +122,18 @@ desc 'Individual test tasks.'
 namespace :test do
   desc 'Test on iOS.'
   task :ios do
-    xcodebuild_in_demo_dir 'build test', :ios, xcprety_args: '--test'
+    xcodebuild 'build test', :ios, xcprety_args: '--test'
     sh "killall Simulator"
   end
 
   desc 'Test on macOS.'
   task :macos do
-    xcodebuild_in_demo_dir 'build test', :macos, xcprety_args: '--test'
+    xcodebuild 'build test', :macos, xcprety_args: '--test'
   end
 
   desc 'Test on tvOS.'
   task :tvos do
-    xcodebuild_in_demo_dir 'build test', :tvos, xcprety_args: '--test'
+    xcodebuild 'build test', :tvos, xcprety_args: '--test'
     sh "killall Simulator"
   end
 
@@ -123,11 +143,11 @@ namespace :test do
     # of the root dir
 
     Dir.mkdir("carthage_test")
-    File.write(File.join("carthage_test", "Cartfile"), "git \"file://#{Dir.pwd}\"")
+    File.write(File.join("carthage_test", "Cartfile"), "git \"file://#{Dir.pwd}\" \"HEAD\"")
     Dir.chdir "carthage_test" do
       sh "carthage bootstrap --platform 'iOS'"
       has_artifacts = Dir.glob("Carthage/Build/*").count > 0
-      raise("Carthage did not succedd") unless has_artifacts
+      raise("Carthage did not succeed") unless has_artifacts
     end
   end
 end
@@ -135,28 +155,23 @@ end
 desc 'Release a version, specified as an argument.'
 task :release, :version do |task, args|
   version = args[:version]
+  release_date = Time.now.strftime("%Y-%m-%d")
   # Needs a X.Y.Z-text format.
-  abort "You must specify a version in semver format." if version.nil? || version.scan(/\d+\.\d+\.\d+/).length == 0
+  abort "You must specify a version in semver format." if version.nil? || version.scan(/\d+\.\d+\.\d+(-\w+\.\d+)?/).length == 0
 
   puts "Updating podspec."
   filename = "Moya.podspec"
   contents = File.read(filename)
-  contents.gsub!(/s\.version\s*=\s"\d+\.\d+\.\d+"/, "s.version      = \"#{version}\"")
+  contents.gsub!(/s\.version\s*=\s"\d+\.\d+\.\d+(-\w+\.\d)?"/, "s.version      = \"#{version}\"")
   File.open(filename, 'w') { |file| file.puts contents }
-
-  puts "Updating Demo project."
-  Dir.chdir('Demo') do
-    ENV['COCOAPODS_DISABLE_DETERMINISTIC_UUIDS'] = 'true'
-    sh "bundle exec pod update Moya"
-  end
 
   puts "Updating changelog."
   changelog_filename = "CHANGELOG.md"
   changelog = File.read(changelog_filename)
-  changelog.gsub!(/# Next/, "# Next\n\n# #{version}")
+  changelog.gsub!(/# Next/, "# Next\n\n# [#{version}] - #{release_date}")
   File.open(changelog_filename, 'w') { |file| file.puts changelog }
 
-  puts "Comitting, tagging, and pushing."
+  puts "Committing, tagging, and pushing."
   message = "Releasing version #{version}."
   sh "git commit -am '#{message}'"
   sh "git tag #{version} -m '#{message}'"
